@@ -4,9 +4,10 @@
   "description": "Integrate with the Tenor GIF API",
   "main": "dist/Tenor.js",
   "scripts": {
+    "compile": "tsc --noEmit",
     "docs": "webpack --mode production",
     "lint": "chq-scripts lint",
-    "prepublishOnly": "rm -rf dist && babel src --out-dir dist --ignore '**/__tests__' && cp src/styles.css dist/styles.css",
+    "prepublishOnly": "rm -rf dist && yarn tsc --declaration && cp src/styles.css dist/styles.css",
     "start": "webpack-dev-server --mode development --hot",
     "test": "chq-scripts test"
   },
@@ -30,7 +31,14 @@
     "@babel/plugin-proposal-class-properties": "^7.5.5",
     "@babel/preset-env": "^7.5.5",
     "@babel/preset-react": "^7.0.0",
-    "@culturehq/scripts": "^3.0.1",
+    "@babel/preset-typescript": "^7.3.3",
+    "@babel/register": "^7.5.5",
+    "@culturehq/scripts": "^4.0.2",
+    "@types/enzyme": "^3.10.3",
+    "@types/jest": "^24.0.18",
+    "@types/react": "^16.9.2",
+    "@types/react-dom": "^16.9.0",
+    "awesome-typescript-loader": "^5.2.1",
     "babel-loader": "^8.0.5",
     "css-loader": "^3.1.0",
     "enzyme": "^3.9.0",
@@ -38,7 +46,8 @@
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
     "style-loader": "^1.0.0",
-    "webpack": "^4.36.1",
+    "typescript": "^3.6.2",
+    "webpack": "^4.39.3",
     "webpack-cli": "^3.3.3",
     "webpack-dev-server": "^3.8.0"
   }
