@@ -1,14 +1,17 @@
-const path = require("path");
+import path from "path";
 
-module.exports = {
+export default {
   output: {
     path: path.join(__dirname, "docs"),
     filename: "index.js"
   },
-  entry: path.join(__dirname, "docs", "app.js"),
+  entry: path.join(__dirname, "docs", "App.tsx"),
+  resolve: {
+    extensions: [".js", ".ts", ".tsx"]
+  },
   module: {
     rules: [
-      { test: /\.js$/, use: "babel-loader", exclude: /node_modules/ },
+      { test: /\.tsx?$/, use: "awesome-typescript-loader" },
       {
         test: /\.css$/,
         use: [{ loader: "style-loader" }, { loader: "css-loader" }],
